Fixed the fscanf DF example by patching `UsageAnalyzer.sameVariable`
