--- joern-cli/frontends/c2cpg/src/test/scala/io/joern/c2cpg/querying/FreeListDataFlowTests.scala
@@ -1,62 +1,0 @@
-package io.joern.c2cpg.querying
-
-import io.joern.c2cpg.testfixtures.DataFlowCodeToCpgSuite
-import io.joern.dataflowengineoss.language._
-import io.shiftleft.semanticcpg.language._
-
-class FreeListDataFlowTests extends DataFlowCodeToCpgSuite {
-
-  override val code: String = """
-    | struct node {
-    |  int value;
-    |  struct node *next;
-    | };
-    |
-    | void free_list(struct node *head) {
-    |  struct node *q;
-    |  for (struct node *p = head; p != NULL; p = q) {
-    |    q = p->next;
-    |    free(p);
-    |  }
-    | }
-    | 
-    | int flow(int p0) {
-    |  int a = p0;
-    |  int b=a;
-    |  int c=0x31;
-    |  int z = b + c;
-    |  z++;
-    |  int x = z;
-    |  return x;
-    | }""".stripMargin
-
-  "should identify all calls to `free`" in {
-    cpg.call.name("free").code.toSetMutable shouldBe Set("free(p)")
-  }
-
-  "should find flows to arguments of `free`" in {
-    implicit val callResolver: NoResolve.type = NoResolve
-    val source = cpg.identifier
-    val sink = cpg.method.name("free").parameter.argument
-    sink.reachableByFlows(source).l.map(flowToResultPairs).distinct.toSet.size shouldBe 5
-  }
-
-  "should find flows to `free`" in {
-    val source = cpg.identifier
-    val sink = cpg.call.name("free")
-    sink.reachableByFlows(source).l.map(flowToResultPairs).distinct.toSet.size shouldBe 5
-  }
-
-  "should find flows from identifiers to return values of `flow`" in {
-    val source = cpg.identifier
-    val sink = cpg.method.name("flow").methodReturn
-    sink.reachableByFlows(source).l.map(flowToResultPairs).distinct.toSet.size shouldBe 9
-  }
-
-  "find flows from z to method returns of flow" in {
-    val source = cpg.identifier.name("z")
-    val sink = cpg.method.name("flow").methodReturn
-    sink.reachableByFlows(source).l.size shouldBe 3
-  }
-
-}
